+++ homework/delete_food.py
@@ -57,5 +57,5 @@ def delete_food(name, amount):
 
 if __name__ == "__main__":
     name = input("제거할 음식 재료를 입력하세요: ")
-    amount = int(input("제거할 수량을 입력하세요: "))
+    amount = float(input("제거할 수량을 입력하세요: "))
     delete_food(name, amount)
